+++ src/extensions/mvs/tree/mvs/mvs-tree-representations.ts
@@ -28,6 +28,11 @@ const Spacefill = {
     ignore_hydrogens: OptionalField(bool, false, 'Controls whether hydrogen atoms are drawn.'),
 };
 
+const Carbohydrate = {
+    /** Scales the corresponding visuals */
+    size_factor: OptionalField(float, 1, 'Scales the corresponding visuals.'),
+};
+
 const Surface = {
     /** Scales the corresponding visuals */
     size_factor: OptionalField(float, 1, 'Scales the corresponding visuals.'),
@@ -42,6 +47,7 @@ export const MVSRepresentationParams = UnionParamsSchema(
         cartoon: SimpleParamsSchema(Cartoon),
         ball_and_stick: SimpleParamsSchema(BallAndStick),
         spacefill: SimpleParamsSchema(Spacefill),
+        carbohydrate: SimpleParamsSchema(Carbohydrate),
         surface: SimpleParamsSchema(Surface),
     },
 );
@@ -52,7 +58,7 @@ const VolumeIsoSurface = {
     /** Absolute isovalue. Overrides `relative_isovalue`. */
     absolute_isovalue: OptionalField(nullable(float), null, 'Absolute isovalue. Overrides `relative_isovalue`.'),
     /** Show mesh wireframe. Defaults to false. */
-    show_wireframe: OptionalField(bool, true, 'Show mesh wireframe. Defaults to false.'),
+    show_wireframe: OptionalField(bool, false, 'Show mesh wireframe. Defaults to false.'),
     /** Show mesh faces. Defaults to true. */
     show_faces: OptionalField(bool, true, 'Show mesh faces. Defaults to true.'),
 };